+++ example/lib/main.dart
@@ -1,16 +1,40 @@
 import 'dart:io';
+import 'dart:ui';
 
 import 'package:flutter/material.dart';
 import 'package:app_center_plugin/app_center_plugin.dart';
-
 
 const androidSecret = '043be909-44a3-4675-b8b3-3078cb55d379';
 const iOSSecret = '894914f3-91d8-48c0-a929-d02e8b5b2178';
 
 void main() async {
-  WidgetsFlutterBinding.ensureInitialized();
   final secret = Platform.isAndroid ? androidSecret : iOSSecret;
+
+  WidgetsFlutterBinding.ensureInitialized();
+
+  // Start AppCenter
   await AppCenter.start(secret);
+
+  // Log Flutter errors with FlutterError.onError
+  FlutterError.onError = (FlutterErrorDetails errorDetails) {
+    FlutterError.presentError(errorDetails);
+    // Send Flutter errors to AppCenter
+    AppCenter.trackError(
+      errorDetails.exceptionAsString(),
+      properties: {"library": errorDetails.library ?? ""},
+      stackTrace: errorDetails.stack,
+    );
+  };
+
+  // Log errors not caught by Flutter with PlatformDispatcher.instance.onError
+  PlatformDispatcher.instance.onError = (error, stack) {
+    AppCenter.trackError(
+      error.toString(),
+      stackTrace: stack,
+    );
+    return true;
+  };
+
   runApp(const MyApp());
 }
 
@@ -45,8 +69,8 @@ class _MyHomePageState extends State<MyHomePage> {
   void _incrementCounter() {
     setState(() {
       _counter++;
-      AppCenter.trackEvent(
-          'Increment Counter', {'counter': _counter.toString()});
+      debugPrint("Sending event to AppCenter");
+      AppCenter.trackEvent('Increment Counter', {'counter': _counter.toString()});
     });
   }
 
@@ -70,11 +94,30 @@ class _MyHomePageState extends State<MyHomePage> {
           ],
         ),
       ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: _incrementCounter,
-        tooltip: 'Increment',
-        child: const Icon(Icons.add),
+      floatingActionButton: Column(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          FloatingActionButton(
+            onPressed: _trackError,
+            tooltip: 'Error',
+            child: const Icon(Icons.bug_report),
+          ),
+          FloatingActionButton(
+            onPressed: _incrementCounter,
+            tooltip: 'Increment',
+            child: const Icon(Icons.add),
+          ),
+        ],
       ),
     );
   }
+
+  void _trackError() {
+    debugPrint("Sending error to AppCenter");
+    throw const CustomException("Custom error message");
+  }
+}
+
+class CustomException implements Exception {
+  const CustomException(String message);
 }