+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [2.0.0] - 2022-10-25
+
+* BREAKING CHANGE: Changed trackError parameters from position parameters to named parameters
+* feat(trackError): Track error's stack trace on Android
+* feat: Upgraded Android target version to 33
+* feat: Upgraded dependencies
+* docs: Updated README usage section to new recommend implementation of catching errors not caught by Flutter
+
 ## [1.0.0] - 2021-11-01
 
 * AppCenter Plugin with Analytics and Crashes.